Fixed #25175 -- Renamed the postgresql_psycopg2 database backend to postgresql.
