+++ django/contrib/gis/db/backends/postgis/base.py
@@ -1,5 +1,5 @@
 from django.db.backends.base.base import NO_DB_ALIAS
-from django.db.backends.postgresql_psycopg2.base import \
+from django.db.backends.postgresql.base import \
     DatabaseWrapper as Psycopg2DatabaseWrapper
 
 from .features import DatabaseFeatures
