+++ django/contrib/gis/db/backends/postgis/features.py
@@ -1,5 +1,5 @@
 from django.contrib.gis.db.backends.base.features import BaseSpatialFeatures
-from django.db.backends.postgresql_psycopg2.features import \
+from django.db.backends.postgresql.features import \
     DatabaseFeatures as Psycopg2DatabaseFeatures
