+++ django/contrib/gis/db/backends/postgis/introspection.py
@@ -1,6 +1,5 @@
 from django.contrib.gis.gdal import OGRGeomType
-from django.db.backends.postgresql_psycopg2.introspection import \
-    DatabaseIntrospection
+from django.db.backends.postgresql.introspection import DatabaseIntrospection
 
 
 class GeoIntrospectionError(Exception):
