+++ django/db/backends/postgresql/base.py
