+++ tests/backends/test_utils.py
@@ -9,7 +9,7 @@ class TestLoadBackend(SimpleTestCase):
         msg = (
             "'foo' isn't an available database backend.\n"
             "Try using 'django.db.backends.XXX', where XXX is one of:\n"
-            "    'mysql', 'oracle', 'postgresql_psycopg2', 'sqlite3'\n"
+            "    'mysql', 'oracle', 'postgresql', 'sqlite3'\n"
             "Error was: No module named %s"
         ) % "foo.base" if six.PY2 else "'foo'"
         with self.assertRaisesMessage(ImproperlyConfigured, msg):
